Remove TODO about supporting multiple documents. For now, let's just support a single document.
Multiple document support is not trivial.   In addition to specifying a document when
creating a widget instance (previously done via new MyWidget({ownerDocument: ...}),
there's document.register() itself, which registers a custom tag with a certain
document.

In the future, multi-document support is needed, maybe it's a question of leveraging
AMD to load multiple instances of the register.js module, each with a different document
specified through module.config().doc.