Pageable: update _setLoading method of PageLoader renderer: update the css class toggling logic.
Pageable: Remove deferred creation when it is not needed, replace use of
when by use of then when possible.

Pageable: Update _idPages documentation.

Pageable: unit test added to verify that when hideOnPageLoad is true,
the loading panel is correctly removed after the last page is loaded and
the page loader is destroyed.