Make Combobox down arrow icon always open the dropdown filtered according to the <input>'s value.
The down arrow key works equivalently to the down arrow icon.
If there's no down arrow icon, then the down arrow key doesn't do anything.

Also, in mobile mode, when the ComboPopup is opened,
copy the value in the Combobox's <input> into the ComboPopup's <input>.

TODO: This commit has known test failures that will be fixed in the next commit.

TODO: Get rid of hasDownArrow property, and instead just show the
down arrow icon iff minFilterChars === 0.  That will simplify the Combobox
logic, or at least the number of scenarios that need to be tested,
because we don't need to check cases like hasDownArrow=true minFilterChars=3.

Refs #668.