Combobox: refactor code for handling click of the <input>.
The old code was overcomplicated and was referencing the non-standard event.srcElement.
It was causing problems on Firefox in a Combobox subclass although for some reason
Combobox itself seemed to work.

Also improved the tests to be explicit about which part of the Combobox they
were clicking: the <input> or the down arrow.