Fix and clean up the features test.
On FF53, the platform name is in environmentType.platformName, not
environmentType.platform.

Also, the testChannelFlagsDone variable didn’t work since it’s
tantamount to a global.  When the test ran on Firefox is didn’t set
testChannelFlagsDone, but the test didn’t fail because that flag had
already been set when the test ran against Chrome.