Replace bizarre programmatic List tests w/something normal.
Also fix grossly inefficient code to count number of items and
categories, which was taking 300s to run on FF53 when the entire test
suite was run.

There was a TODO in the code to check the label of each
item, but even if we do that we shouldn't be passing back entire
Elements.