Combobox: switch between centered dialog vs. dropdown based on viewport size when popup opened.
Note that the code sharing between Combobox and ComboPopup is still very
messy IMO.  I think the ComboPopup and Combobox should extend the same
base class, rather than ComboPopup referencing this.combobox.list and
this.combobox._prepareInput().  In other words, Combobox should instantiate
a ComboPopup, but ComboPopup should be an independent class.  However,
that’s not addressed in this check in.

Note also that both before and after this change, keyboard navigation
(up/down arrows) doesn’t work for a ComboPopup with filtering <input>.