Remove autoPaging test.
It was already disabled for Firefox, and was failing for Chrome and
Edge (albeit in a different way).  Plus, my manual tests of
pageable-prog-6.html show inconsistent behavior.  Sometimes when
a new page loads it will scroll to the bottom of the new page,
whereas other times it will keep scroll position where it was before.

And finally, I have no idea what the title of the test means about
"category headers destroyed when loading pages".