Let browser handle tabbing out of List.
The code to tab out of a list programatically was too problematic,
specifically where the List was the last element on the page.
In that case the focus should go to the browser's address bar etc.,
which is not possible to do programatically.

Note that the same issue existed for shift-tab when the List was the
first element on the page.

Refs #721.