Implement filterable multiselect.
Previously this wasn't possible because multiselect moved focus to the dropdown
to avoid some JAWS issue.  But that doesn't seem to be a problem anymore.

The user navigates between the options using the arrow keys, toggles options using
the ENTER key, and closes the dropdown using the ESC key.
Like single-select, normal characters and the space character are used for searching.

When filtering is enabled, there is no way to see which items are selected,
or even how many are selected, without opening the dropdown.  In other words,
the <input> never contains anything except a search string the user typed.
A subclass can add an indicator to the template to show how many items are selected.

Other changes include:

* Rename searchPlaceHolder to placeholder.  Unclear why it previously had a special name.
* Fix Combobox left/right arrow keys when dropdown open.
  The issue was that HasDropDown forwards left and right arrows to the dropdown,
  and then the event gets canceled.
* Prevent spurious input event on IE from leaving Combobox.
* Remove vestigial properties, and properties already inherited from LitFormWidget and
  LitFormValueWidget.
* Update old API documentation.